Функции игровой логики вынесены в отдельный модуль
