+++ js/level-genre.js
@@ -1,4 +1,5 @@
-import {getElementFromTemplate, getQuestion, showNextScreen} from './utils';
+import {getElementFromTemplate} from './utils';
+import {getQuestion, showNextScreen} from './data/game-logic';
 import renderHeader from './game/header';
 import renderPlayer from './game/player';
