+++ js/utils.js
@@ -1,9 +1,3 @@
-import {GAME} from './data/game-data';
-import genreLevel from './level-genre';
-import artistLevel from './level-artist';
-import winResult from './win-result.js';
-import failResult from './fail-result';
-
 export const getElementFromTemplate = (template) => {
   const outerElement = document.createElement(`div`);
   outerElement.innerHTML = template;
@@ -37,35 +31,4 @@ export const getMinuteAndSeconds = (time) => {
   const seconds = `${Math.floor(time % 60) < 10 ? `0` : ``}${Math.floor(time % 60)}`;
 
   return {minutes, seconds};
-};
-
-export const getQuestion = (state) => state.questions[state.currentQuestion];
-
-export const getQuestionView = (state) => {
-  const question = getQuestion(state);
-  let questionView;
-  if (question.type === `artist`) {
-    questionView = artistLevel(state);
-  } else if (question.type === `genre`) {
-    questionView = genreLevel(state);
-  }
-  return questionView;
-};
-
-const canContinue = (state) => state.lives > 0;
-
-const win = (state) => canContinue(state) && state.currentQuestion === GAME.MAX_QUESTIONS;
-
-export const showNextScreen = (state) => {
-  let nextScreen;
-  if (!canContinue(state)) {
-    nextScreen = failResult(state);
-  } else if (win(state)) {
-    nextScreen = winResult(state);
-  }
-  if (nextScreen) {
-    showView(nextScreen);
-  } else {
-    showView(getQuestionView(state));
-  }
 };
